requirements.txt
Flask==2.2.3
keras==2.7.0
numpy==1.22.4
opencv_python==4.5.5.62
Pillow==10.1.0
Werkzeug==2.2.3
